bug fix in user activity pt 2

+++ app/src/main/java/com/cmput301w18t26/taskit/UserActivity.java
@@ -51,7 +51,7 @@ public class UserActivity extends AppCompatActivity {
                     String usernameInput = usernameEdit.getText().toString();
                     String nameInput = nameEdit.getText().toString();
                     String emailInput = emailEdit.getText().toString();
-                    long phoneInput = Integer.parseInt(phoneEdit.getText().toString());
+                    long phoneInput = Long.parseLong(phoneEdit.getText().toString());
                     if (db.userExists(usernameInput)) {
                         invalidUsername.setText("Username already exists");
                         invalidUsername.setVisibility(View.VISIBLE);
